resolve: merge interfaces and impls
