--- base/src/main/java/org/aya/resolve/context/MutableModuleSymbol.java
@@ -1,131 +1,0 @@
-// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
-// Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.resolve.context;
-
-import kala.collection.Map;
-import kala.collection.MapView;
-import kala.collection.SetView;
-import kala.collection.mutable.MutableMap;
-import kala.control.Option;
-import kala.control.Result;
-import kala.tuple.Tuple2;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.function.BiConsumer;
-
-/**
- * A container of symbols.
- *
- * @param table `Unqualified -> (Component -> Symbol)`<br/>
- *              It says a `Symbol` can be referred by `{Component}::{Unqualified}`
- * @apiNote the methods that end with `Definitely` will get/remove only one symbol or fail if ambiguous.
- */
-public record MutableModuleSymbol<T>(
-  @NotNull MutableMap<String, MutableMap<ModulePath, T>> table
-) implements ModuleSymbol<T> {
-  public MutableModuleSymbol() {
-    this(MutableMap.create());
-  }
-
-  public MutableModuleSymbol(@NotNull ModuleSymbol<T> other) {
-    this(other.toMap().toImmutableSeq()
-      .collect(MutableMap.collector(
-        Tuple2::component1,
-        x -> MutableMap.from(x.component2()))
-      )
-    );
-  }
-
-  @Override
-  public @NotNull Map<String, Map<ModulePath, T>> toMap() {
-    return Map.from(table().toImmutableSeq());
-  }
-
-  @Override
-  public @NotNull SetView<String> keysView() {
-    return table().keysView();
-  }
-
-  @Override
-  public @NotNull MapView<String, Map<ModulePath, T>> view() {
-    return table().view().mapValues((k, v) -> v);
-  }
-
-  /**
-   * Adding a new symbol which can be referred by `{componentName}::{name}`
-   */
-  public Option<T> add(
-    @NotNull ModulePath componentName,
-    @NotNull String name,
-    @NotNull T ref
-  ) {
-    var candidates = table().getOrPut(name, MutableMap::create);
-    return candidates.putIfAbsent(componentName, ref);
-  }
-
-  public void addAnyway(
-    @NotNull ModulePath componentName,
-    @NotNull String name,
-    @NotNull T ref
-  ) {
-    this.table().getOrPut(name, MutableMap::create).put(componentName, ref);
-  }
-
-  @Override
-  public @NotNull Map<ModulePath, T> resolveUnqualified(@NotNull String unqualifiedName) {
-    var result = table().getOption(unqualifiedName);
-    if (result.isEmpty()) {
-      return Map.empty();
-    } else {
-      return result.get();
-    }
-  }
-
-  public @NotNull MutableMap<ModulePath, T> getMut(@NotNull String unqualifiedName) {
-    return table().getOrPut(unqualifiedName, MutableMap::create);
-  }
-
-  public Option<T> remove(@NotNull ModulePath component, @NotNull String unqualifiedName) {
-    return table().getOption(unqualifiedName).flatMap(x -> x.remove(component));
-  }
-
-  public Option<Map<ModulePath, T>> removeAll(@NotNull String unqualifiedName) {
-    return Option.narrow(table().remove(unqualifiedName));
-  }
-
-  public Result<T, Error> removeDefinitely(@NotNull String unqualifiedName) {
-    var candidates = getMut(unqualifiedName);
-
-    if (candidates.isEmpty()) return Result.err(Error.NotFound);
-    if (candidates.size() != 1) return Result.err(Error.Ambiguous);
-
-    var result = candidates.iterator().next().getValue();
-    candidates.clear();
-
-    return Result.ok(result);
-  }
-
-  public Result<T, Error> removeDefinitely(@NotNull ModulePath component, @NotNull String unqualifiedName) {
-    return switch (component) {
-      case ModulePath.Qualified qualified -> {
-        var result = remove(component, unqualifiedName);
-
-        if (result.isEmpty()) {
-          yield Result.err(Error.NotFound);
-        } else {
-          yield Result.ok(result.get());
-        }
-      }
-      case ModulePath.This aThis -> removeDefinitely(unqualifiedName);
-    };
-  }
-
-  @Override
-  public void forEach(@NotNull BiConsumer<String, Map<ModulePath, T>> action) {
-    table().forEach(action);
-  }
-
-  public void clear() {
-    table().clear();
-  }
-}
